pocket_robot/requirements.txt
pydantic
psutil
fastapi
uvicorn
python-socketio
aiofiles
